Update join_the_odin_community.md
When I first attempted to link my GitHub to my Discord profile, I was first confused as to whether the Settings mentioned I was supposed to connect through was in Discord or GitHub.  The instructions weren't precisely clear enough for me.  I had to explore the settings in both GitHub and Discord to finally find it.  When I finally figured out the settings to be modified were in Discord, initially, there was no icon for GitHub.  I watched a Youtube video and learned you needed to be both logged in and have Discord and GitHub open to have the GitHub icon appear in Discord Settings > Connections.